Round screen width/height up to next multiple of 4
Mako88 reports incorrect images captured when using a screen resolution
of 1366x768.  1366 is not a multiple of 4.  Debugging also shows memory
corruption at this resolution, as Mesa writes off the end of the
allocated buffer.  Padding the buffer to tolerate these writes is
insufficient, as libpng then crashes with an alignment fault trying to
read unaligned data from the buffer.  All these problems are eliminated
by rounding the width and height to the next multiple of 4.

Reported-by: Mako88 <https://github.com/dxx-rebirth/dxx-rebirth/issues/373>
Fixes: 131c1b9f4d556fbd9d67815755608602b408698b ("Add support for PNG screenshots")